Packages were restructured to improve dependencies and mock generation.
