+++ Modules/feature-test/Sources/Utils/Constants.swift
@@ -36,22 +36,3 @@ extension Constants {
 extension Constants {
   static let randomIdentifier = UUID().uuidString
 }
-
-
-extension Constants {
-  struct MockPresentationService: PresentationService {
-    
-    var flow: EudiWalletKit.FlowType
-    
-    func startQrEngagement() async throws -> Data? { nil }
-    
-    func receiveRequest() async throws -> [String : Any] { [:] }
-    
-    func sendResponse(userAccepted: Bool, itemsToSend: EudiWalletKit.RequestItems, onSuccess: ((URL?) -> Void)?) async throws {}
-  }
-  
-  static let mockPresentationSession = PresentationSession(
-    presentationService: MockPresentationService(flow: .other),
-    userAuthenticationRequired: false
-  )
-}
