+++ Modules/logic-business/Package.swift
@@ -37,7 +37,6 @@ let package = Package(
       url: "https://github.com/rhummelmose/BluetoothKit.git",
       branch: "master"
     ),
-    .package(name: "logic-core", path: "./logic-core"),
     .package(name: "logic-test", path: "./logic-test"),
     .package(name: "logic-analytics", path: "./logic-analytics")
   ],
@@ -50,7 +49,6 @@ let package = Package(
         "IOSSecuritySuite",
         "KeychainAccess",
         "BluetoothKit",
-        "logic-core",
         .product(
           name: "Peppermint",
           package: "peppermint"
